clear some files, modified class definition

+++ Assets/Script/Inventory/Gems.cs
@@ -6,7 +6,7 @@ public class Gems : Item
 {
     public GemType gType { get; set; }
     public ushort Skill { get; set; }
-    public Gems(int id, string name, string description, int capacity, ItemType type, double buyprice, string sprite, GemType gtype, ushort skill)
+    public Gems(int id, string name, string description, int capacity, ItemType type, int buyprice, string sprite, GemType gtype, ushort skill)
         : base(id, name, description, capacity, type, buyprice, sprite)
     {
         gType = gtype;
